# perf: speed up overlay performance ~200-40000x depending on the size of the document